backend/pyproject.toml
[project]
name = "botc-bench"
version = "0.1.0"
description = "Blood on the Clocktower AI Agent Benchmark"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.52.0",
    "openai>=1.75.0",
    "google-genai>=1.14.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["botc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
